pyproject.toml
[tool.poetry]
name = "pd-server"
version = "0.1.0"
description = ""
authors = ["ffenland <ffenland@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
firebase-admin = "^6.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
